+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- `Select` and `SelectMulti` option groups `title` is now `label` and optional.
 - `Badge` style updated to use lighter colors for intents. Badges are now always round.
 
 ### Fixed